chore: improve authorize page
