+++ packages/mock-server/src/createMockServer.ts
@@ -3,8 +3,8 @@ import type { OpenAPI, OpenAPIV3_1 } from '@scalar/openapi-types'
 import { type Context, Hono } from 'hono'
 import { cors } from 'hono/cors'
 
-import { respondWithAuthorizePage } from './respondWithAuthorizePage'
 import { mockAnyResponse } from './routes/mockAnyResponse'
+import { respondWithAuthorizePage } from './routes/respondWithAuthorizePage'
 import { respondWithOpenApiDocument } from './routes/respondWithOpenApiDocument'
 import type { HttpMethod, MockServerOptions } from './types'
 // import { anyBasicAuthentication } from './utils/anyBasicAuthentication'
